Move the files to functional_components folder. Update the imported files from absolute path.
